Fix clippy warnings
This commit is a fairly mechanical fixing of all the clippy warnings in
the repository. Most of the time the clippy suggestions are just
directly applied. Some of the issues clippy pointed out probably could
benefit from a deeper refactoring around some of the types and function
signatures to reduce copying, owned types, and allocations necessary,
but that can be saved for a follow up PR. In a follow up commit we
should add a ci job that runs `cargo fmt --check` and `cargo clippy` to
enforce the style and lint as part of pre-merge checking.