Fix clippy failure on latest nightly
In newer nightly and the current rust beta channel there is a new clippy
rule that suggests wrapping mixed bitshift and bitwise operators in
parenthesis to make the precedence between the operators more explicit.
While the rules seem clear to me, this is failing CI and simple to
address, so this commit adds the recommend parenthesis to fix the
failing job.